feat!: Drop C and Rust APIs
The C and Rust APIs were proof-of-concept initiatives, to demonstrate
the simplicity of a certain foreign language usage of Qibolab.

This is not the way we want to eventually support multi-language API,
which will be instead be based on `qibo-core`, or directly exposing a
multi-language API for Qibolab itself, just making calls to Python-based
drivers (since most vendor APIs are provided in Python).

The latter is becoming increasingly simple by the standardization and
stabilization of the drivers' interface, which could rely on easily
serialized inputs and outputs.

To avoid dissipating workforce in maintaining these proof-of-concept
elements, which have little to no practical usage, I'm phasing out them
immediately, since they would need some fixes anyhow.